Improve configuration for releasing artifacts
The `sbt-sonatype` plugin was uploading the artifacts, but not closing or
releasing the artifact-repository. I think this is because the
`sonatypeProfileName` was not correctly configured:

```
[info] sonatypeProfileName = root
[info] Reading staging repository profiles...
log4j:WARN No appenders could be found for logger (org.apache.http.impl.conn.BasicClientConnectionManager).
log4j:WARN Please initialize the log4j system properly.
[warn] No staging repository is found. Do publishSigned first.
```

I've also attempted to disable publishing of the 'root' project artifact.