Keep test console output when junit output is used
Before now, scalatest’s output was never sent to standard out: the
default behaviour when passing the “-u” flag is to omit it. This is a
bit irritating when running locally, because it makes it hard to tell
which tests have failed.

This commit changes the behaviour to always send scalatest’s output to
standard out, as well as writing to junit files.

This lets us keep the benefits of the XML output in CI (annotations
pointing to the source files), while making the tests more useful when
run locally.